+++ packages/codegen.go/src/metadata.ts
@@ -6,7 +6,7 @@
 import * as go from '../../codemodel.go/src/index.js';
 
 // Creates the content in _metadata.json
-export async function generateMetadataFile(codeModel: go.CodeModel): Promise<string> {
+export function generateMetadataFile(codeModel: go.CodeModel): string {
   // Return the formatted JSON string
   return JSON.stringify(codeModel.metadata, null, 2);
 }